+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add guidance on setting label and legend sizes
 
+:wrench: **Maintenance**
+
+- Use NHS.UK frontend CSS and JS in design example previews
+
 ## 8.5.0 - 27 November 2025
 
 :new: **New features**